update mapa and add pages
