+++ src/app/(pages)/analizar/page.tsx
@@ -1,0 +1,8 @@
+
+export default function AnalizarPage() {
+  return (
+    <div>
+      <h1>Hello Page</h1>
+    </div>
+  );
+}
