+++ src/app/layout.tsx
@@ -27,7 +27,6 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <link href="https://api.mapbox.com/mapbox-gl-js/v3.1.2/mapbox-gl.css" rel="stylesheet"  />
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
